something working

+++ src/controllers/DoController.php
@@ -68,15 +68,20 @@ class DoController extends \yii\console\Controller
             die('Name!');
         }
 
+        $full_name = $type . '-asset/' . $name;
         $regurl = $apiurl . $name;
         $data = Json::decode($this->getContents($regurl, $regurl));
 
         $url = $data['url'];
         $repo = new AssetVcsRepository(['url' => $url, 'type' => $type . '-assets'], $this->io, $this->composer->getConfig(), $this->composer->getEventDispatcher());
-        $packages = [];
-        $full_name = $type . '-asset/' . $name;
+        $versions = [];
         foreach ($repo->getPackages() as $package) {
-            $packages[$package->getVersion()] = [
+            if (!$package->getDistType()) {
+                /// TODO investigate more why
+                continue;
+            }
+            $versions[$package->getVersion()] = [
+                'uid'       => rand(1000000, 2000000),
                 'name'      => $full_name,
                 'version'   => $package->getVersion(),
                 'dist'      => [
@@ -85,15 +90,73 @@ class DoController extends \yii\console\Controller
                 ],
             ];
         }
-        $json = Json::encode($packages);
-        var_dump($json);
+
+        return $this->updatePackage($full_name, $versions);
+    }
+
+    static public function mkdir($dir)
+    {
+        if (!file_exists($dir)) {
+            mkdir($dir, 0777, true);
+        }
+    }
+
+    protected function updatePackage($name, $versions)
+    {
+        $data = [
+            'packages' => [
+                $name => $versions
+            ],
+        ];
+        $json = Json::encode($data);
         $hash = hash('sha256', $json);
-        var_dump($hash);
-        $path = $this->getPath($name, $hash);
-        var_dump($path);
-        file_put_contents($path, $json);
+        $path = Yii::getAlias("@web/p/$name$$hash.json");
+        if (!file_exists($path)) {
+            static::mkdir(dirname($path));
+            file_put_contents($path, $json);
+            $this->updateProviderLatest($name, $hash);
+        }
 
         return $hash;
+    }
+
+    protected function updateProviderLatest($name, $hash)
+    {
+        $latest_path = Yii::getAlias('@web/provider-latest.json');
+        $data = Json::decode(file_get_contents($latest_path) ?: '[]');
+        if (!is_array($data)) {
+            $data = [];
+        }
+        if (!isset($data['providers'])) {
+            $data['providers'] = [];
+        }
+        $data['providers'][$name] = ['sha256' => $hash];
+        $json = Json::encode($data);
+        $hash = hash('sha256', $json);
+        $path = Yii::getAlias("@web/p/provider-latest$$hash.json");
+        if (!file_exists($path)) {
+            file_put_contents($path, $json);
+            /// TODO lock
+            file_put_contents($latest_path, Json::encode($data));
+            $this->updateIndex($hash);
+        }
+
+        return $hash;
+    }
+
+    protected function updateIndex($hash)
+    {
+        $data = [
+            'providers-url'     => '/p/%package%$%hash%.json',
+            'search'            => '/search.json?q=%query%',
+            'provider-includes' => [
+                'p/provider-latest$%hash%.json' => [
+                    'sha256' => $hash,
+                ],
+            ],
+        ];
+        /// TODO lock
+        file_put_contents(Yii::getAlias('@web/packages.json'), Json::encode($data));
     }
 
     public function getPath($type, $name, $hash)
@@ -104,5 +167,10 @@ class DoController extends \yii\console\Controller
             mkdir($dir, 0777, true);
         }
         return $dir . "/$hash.json";
+    }
+
+    public function actionTest()
+    {
+        $this->updateProviderLatest('sadfasd/sadfa');
     }
 }
